fix(ui): move the srv/client identity rail into a fixed status footer
The rail lived in the tab strip's single flex-wrap row holding its
ml-auto; once host tabs filled the width it wrapped onto an orphaned,
mostly-empty second row. A slim viewport-fixed bottom bar (the VS Code
/ tmux placement) can never collide with the tabs, survives page
scroll, and absorbs the phone home-indicator inset for the installed-
PWA case. The theme toggle inherits the strip's ml-auto.
