+++ README.md
@@ -41,7 +41,7 @@ Open <http://localhost:7720>. The UI opens on the **fleet** tab — a single ove
 - **Idle NICs collapse** behind a `+N idle` toggle by default, so the few interfaces moving traffic aren't buried under the dozens of always-zero virtual ones (utunN, anpiN, …).
 - **Strictly read-only** — drishti only ever *observes* a host. The per-host surface exposes no procedures, so there is no way to signal, kill, or otherwise mutate a monitored process through the UI or the wire.
 - **Installable PWA, one per host** — drishti ships a web app manifest and an emerald aperture icon, so you can install it as a standalone app (desktop, or a phone's home screen). The manifest's name and identity are the server's own `drishti@<hostname>`, so installing drishti from two different hosts gives you **two distinct, separately-labelled apps** in the OS app list — not one ambiguous `drishti` that collides. The shell and its assets are served by [`@kolu/surface-app`](https://github.com/juspay/kolu/tree/master/packages/surface-app) under a strict freshness contract (see below); drishti ships **no caching service worker** of its own. A **Pin app** button in the tab strip surfaces the install prompt directly — it appears only on a secure origin (`https://…`, e.g. a `tailscale serve` FQDN) and hides once drishti is already installed, gated on the `canInstallPwa` / `isInstalled` signals from `@kolu/surface-app`'s `useSurfaceApp()` and rendered via [`@kolu/solid-pwa-install`](https://github.com/juspay/kolu/tree/master/packages/solid-pwa-install) (which owns the cross-browser install volatility). <sub>The button is gated behind `TODO(pin)` until the kolu pin tracks the in-flight `welcome` revision that ships those signals + that package — see the draft-PR note below.</sub>
-- **Always the deployed build** — drishti adopts `@kolu/surface-app`, which owns the *freshness contract* on the wire: the HTML shell is served `no-store`, the content-hashed `/assets/*` bundle and stylesheet are pinned `immutable` for a year, an asset miss 404s (never the HTML shell under a `.js` URL), and the server serves a self-destructing `/sw.js` that unregisters any legacy caching worker an earlier drishti build left behind. The client's build commit (baked into the bundle as `__SURFACE_APP_COMMIT__`) rides a `buildInfo` cell on the admin surface alongside the server's resolved commit; when a tab is provably behind the deployed build the tab strip shows a `≠ server — reload` affordance. A returning installed client thus always re-fetches the shell and converges to the deployed build — the stale-client class of bug, gone structurally.
+- **Always the deployed build** — drishti adopts `@kolu/surface-app`, which owns the *freshness contract* on the wire: the HTML shell is served `no-store`, the content-hashed `/assets/*` bundle and stylesheet are pinned `immutable` for a year, an asset miss 404s (never the HTML shell under a `.js` URL), and the server serves a self-destructing `/sw.js` that unregisters any legacy caching worker an earlier drishti build left behind. The client's build commit (baked into the bundle as `__SURFACE_APP_COMMIT__`) rides a `buildInfo` cell on the admin surface alongside the server's resolved commit; both commits — plus a server-connection liveness dot — are always visible in a slim status footer pinned to the bottom of the viewport, and when a tab is provably behind the deployed build the footer grows a `≠ srv` one-tap reload affordance. A returning installed client thus always re-fetches the shell and converges to the deployed build — the stale-client class of bug, gone structurally.
 
 Requirements:
